tests refactoring
